Initial implementation of the Nginx HSTS module
The initial design limits HSTS configuration to main and server, but not to location. The reason behind is that the scope of HSTS protection is on the entire virtual host instead of a particular location. So if the intention is to protect a virtual host using HSTS, it is a best practice to reply the HSTS header for every request to the site.

Sample configuration:
hsts on
hsts_max_age 31536000
hsts_includesubdomains on
hsts_preload on